Split sidebar surface contracts
