+++ Tests/QuillCodeAppTests/QuillCodeThreadSidebarSurfaceTests.swift
@@ -2,48 +2,7 @@ import XCTest
 import QuillCodeCore
 @testable import QuillCodeApp
 
-final class QuillCodeSidebarSurfaceTests: XCTestCase {
-    func testProjectItemSurfaceBuildsRemoteStateAndDefaultActions() {
-        let projectID = UUID()
-        let connection = ProjectConnection.ssh(
-            path: "/srv/quill",
-            host: "feather.local",
-            user: "quill",
-            port: 22
-        )
-        let project = ProjectRef(id: projectID, name: "Feather", path: connection.path, connection: connection)
-        let item = ProjectItemSurface(project: project, selectedProjectID: projectID)
-
-        XCTAssertEqual(item.id, projectID)
-        XCTAssertEqual(item.name, "Feather")
-        XCTAssertEqual(item.path, "ssh://quill@feather.local:22/srv/quill")
-        XCTAssertEqual(item.connectionKindLabel, "SSH Remote")
-        XCTAssertTrue(item.isRemote)
-        XCTAssertTrue(item.isSelected)
-        XCTAssertEqual(item.actions.map(\.kind), [.newChat, .refreshContext, .rename, .remove])
-        XCTAssertEqual(item.actions.map(\.kind.title), ["New chat", "Refresh context", "Rename", "Remove from list"])
-        XCTAssertEqual(item.actions.first?.id, "\(projectID.uuidString)-newChat")
-    }
-
-    func testProjectItemSurfaceDecodesOlderPayloadWithoutRemoteMetadataOrActions() throws {
-        let projectID = UUID()
-        let json = """
-        {
-          "id": "\(projectID.uuidString)",
-          "name": "QuillCode",
-          "path": "/Users/quill/QuillCode",
-          "isSelected": false
-        }
-        """
-        let data = try XCTUnwrap(json.data(using: .utf8))
-        let item = try JSONDecoder().decode(ProjectItemSurface.self, from: data)
-
-        XCTAssertEqual(item.connectionKindLabel, "Local")
-        XCTAssertFalse(item.isRemote)
-        XCTAssertEqual(item.actions.map(\.kind), [.newChat, .refreshContext, .rename, .remove])
-        XCTAssertTrue(item.actions.allSatisfy(\.isEnabled))
-    }
-
+final class QuillCodeThreadSidebarSurfaceTests: XCTestCase {
     func testSidebarSurfaceFiltersGroupsAndLabelsSelection() {
         let selectedThread = ChatThread(title: "Run whoami", model: TrustedRouterDefaults.synthModel)
         var pinnedThread = ChatThread(title: "Review git diff", model: "z-ai/glm-5.2", isPinned: true)
